Add files via upload
Python programs based on regression, Tree and SVM models in order to differentiate between malicious and benign URL'S and represent the same in the form of a dynamic graph. The input.csv file is to be replaced with an input file of your choice. The code initially builds it's knowledge base by taking into account a number of URL'S and observing it's characteristics and it's corresponding nature. When this knowledge base is complete, this program uses the same knowledge base and rules to predict the nature of any other URL'S supplied via the input.csv file. The prediction is also represented in the form of an automated graph.